join, login MainController로 이동, UserController API 수정

+++ src/main/java/com/example/runners/dto/challenge/ChallengeDTO.java
@@ -13,6 +13,6 @@ public class ChallengeDTO {
     private String description;
     private int goalAmount;
     private String goalUnit;
-    private int createdBy; // 작성자
+    private long createdBy; // 작성자
     private Date deadline;
 }
